Intercept number parsing error from the parser and throw a C++ exception.
CMakeLists.txt: Complied Bison and flex files with C++
SetVar.cc: VerifyVal() now intercepts number parsing errors also.
clError.h: Moved clBoolParsingError type here, and removed boolError.h
callback.cc, clParseVal.cc, isBool.cc: Removed inclusion of boolError.h
